spawner/requirements.txt
azure-common==1.1.27
azure-mgmt-resource==20.0.0
azure-mgmt-containerinstance==9.1.0
azure.servicebus